+++ backend/package.json
@@ -25,7 +25,6 @@
     "dotenv": "^8.2.0",
     "eslint-plugin-promise": "^4.2.1",
     "express": "~4.16.1",
-    "express-session": "^1.17.0",
     "express-ws": "^4.0.0",
     "lodash": "^4.17.15",
     "morgan": "~1.9.1",
