--- backend/src/domain/login/estaLogueado.js
@@ -1,5 +1,0 @@
-export default function estaLogueado(req){
-    return req && req.session
-    && req.session.usuario 
-    && req.session.usuario.root;
-}
