--- backend/src/domain/perfil/controller.js
@@ -1,11 +1,0 @@
-import { pick } from 'lodash';
-
-const Controller = () => ({
-
-    me: (req) => {
-        return pick(req.session.usuario, 'nombre')
-    },
-
-});
-
-export default Controller;
