+++ backend/src/domain/pruebas/pruebas.test.js
@@ -5,5 +5,5 @@ describe('Post Endpoints', () => {
   it('should create a new post', () => request(app)
     .get('/api/pruebas/temas')
     .send()
-    .then((res) => expect(res.statusCode).toEqual(403)));
+    .then((res) => expect(res.statusCode).toEqual(200)));
 });
