+++ backend/src/server.js
@@ -1,7 +1,7 @@
 import express, { json, urlencoded } from 'express';
+import cookieParser from 'cookie-parser';
 import morgan from 'morgan';
 import * as path from 'path';
-import session from 'express-session';
 
 import indexRouter from './routes';
 import webSocketRouter from './webSocket';
@@ -14,8 +14,7 @@ const expressWs = require('express-ws')(app);
 app.use(morgan('combined', { stream: logger.stream }));
 app.use(json());
 app.use(urlencoded({ extended: false }));
-// TODO: Express sessions esta usando un memory store, deberiamos mandar en la cookie directamente el valor (a la JWT) sino cada reset de heroku invalida las sesiones anteriores.
-app.use(session({ secret: 'keyboard cat' }))
+app.use(cookieParser());
 
 app.ws('/ws', webSocketRouter(expressWs.getWss()));
 app.use('/api', indexRouter);
