--- frontend/src/Authenticator.js
@@ -1,40 +1,0 @@
-import React from 'react';
-import { createBrowserHistory } from 'history';
-import { Route, Router } from 'react-router-dom';
-import App from './App';
-import Login from './login/Login';
-import Backend from './api/backend';
-import Oradores from './oradores';
-import Mobile from './mobile';
-import TestChart from './chart';
-
-const history = createBrowserHistory();
-
-export default class Authenticator extends React.Component {
-  state = {
-    usuario: null,
-    cargando: true,
-  }
-
-  componentDidMount() {
-    Backend.getPerfil()
-      .then((usuario) => this.setState({ usuario }))
-      .catch((error) => error.response.status !== 403 && Promise.reject(error))
-      .finally(() => this.setState({ cargando: false }));
-  }
-
-  render() {
-    const { cargando, usuario } = this.state;
-
-    if (cargando) return <h1>Cargando</h1>;
-    if (!usuario) return <Login />;
-    return (
-      <Router history={history}>
-        <Route path="/" render={({ location }) => <App location={location}/>}/>
-        <Route path="/mobile" render={({ location }) => <Mobile location={location}/>}/>
-        <Route path="/oradores" render={({ location }) => <Oradores location={location}/>}/>
-        <Route path="/chart" render={({ location }) => <TestChart location={location}/>}/>
-      </Router>
-    );
-  }
-}
