+++ frontend/src/api/backend.js
@@ -22,10 +22,6 @@ const Backend = {
   getTemas() {
     return requester.get('/temas/obtener');
   },
-
-  getPerfil() {
-    return requester.get('/perfil/me');
-  },
 };
 
 export default Backend;
