+++ frontend/src/reunion/TemasHandler.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import { toast } from 'react-toastify';
 import backend from '../api/backend';
 import VistaTemas from './VistaTemas';
+import { toast } from 'react-toastify';
 
 class TemasHandler extends React.Component {
+
   constructor(props) {
     super(props);
     this.state = {
@@ -46,6 +47,7 @@ class TemasHandler extends React.Component {
       default: return null;
     }
   }
+
 }
 
 export default TemasHandler;
